kafka_producer/requirements.txt
Faker==18.9.0
kafka-python==2.0.2
schedule==1.2.0
